fix issue related to a failed ConnectionRequest
Whenever a user tries to connect to a non existent host,
the nio connect() call would fail and throw an exception but will
also trigger an nio event that in turns throws an exception on
finishConnect(). Both exception would be propagated to the rx subscriber
however since the first one would have closed the subscription,
the second one would failed to be deliver and throws a runtime rx
exception.

The fix was simply to check whether the subscriber was disposed or not.

Resolves: #1